Remove applying pet effect on `ItemEnhancement`
