+++ .Lib9c.Tests/Action/Scenario/Pet/ReduceRequiredBlockTest.cs
@@ -127,105 +127,5 @@ namespace Lib9c.Tests.Action.Scenario.Pet
             // TEST: RequiredBlockIndex
             Assert.Equal(expectedBlock, slotState.RequiredBlockIndex);
         }
-
-        [Theory]
-        [InlineData(0, 10114000, null, "success")]
-        [InlineData(0, 10114000, 1, "success")] // Lv.1 reduces 5.5%
-        [InlineData(0, 10114000, 30, "success")] // Lv.30 reduces 20%
-        [InlineData(14, 10114000, null, "greatSuccess")]
-        [InlineData(14, 10114000, 1, "greatSuccess")] // Lv.1 reduces 5.5%
-        [InlineData(14, 10114000, 30, "greatSuccess")] // Lv.30 reduces 20%
-        [InlineData(10, 10114000, null, "fail")]
-        [InlineData(10, 10114000, 1, "fail")] // Lv.1 reduces 5.5%
-        [InlineData(10, 10114000, 30, "fail")] // Lv.30 reduces 20%
-        public void ItemEnhancementTest(
-            int randomSeed,
-            int targetItemId,
-            int? petLevel,
-            string resultType
-        )
-        {
-            const int itemLevel = 4;
-            var avatarState = _initialStateV2.GetAvatarStateV2(_avatarAddr);
-            var random = new TestRandom(randomSeed);
-            var equipmentRow =
-                _tableSheets.EquipmentItemSheet.Values.First(e => e.Id == targetItemId);
-            var equipment = (Equipment)ItemFactory.CreateItemUsable(
-                equipmentRow,
-                default,
-                0,
-                itemLevel
-            );
-            var material = (Equipment)ItemFactory.CreateItemUsable(
-                equipmentRow,
-                Guid.NewGuid(),
-                0,
-                itemLevel
-            );
-            var enhancementRow = _tableSheets.EnhancementCostSheetV2.Values.First(
-                cost => cost.ItemSubType == ItemSubType.Weapon
-                        && cost.Grade == equipment.Grade
-                        && cost.Level == equipment.level + 1
-            );
-            var expectedBlock = resultType switch
-            {
-                "success" => enhancementRow.SuccessRequiredBlockIndex,
-                "greatSuccess" => enhancementRow.GreatSuccessRequiredBlockIndex,
-                _ => enhancementRow.FailRequiredBlockIndex
-            };
-
-            // Give equipments
-            avatarState.inventory.AddItem(equipment);
-            avatarState.inventory.AddItem(material);
-            var stateV2 =
-                _initialStateV2.SetState(_inventoryAddr, avatarState.inventory.Serialize());
-
-            // Get pet
-            if (!(petLevel is null))
-            {
-                var petRow = _tableSheets.PetOptionSheet.Values.First(
-                    pet => pet.LevelOptionMap[(int)petLevel].OptionType == PetOptionType
-                );
-                _petId = petRow.PetId;
-                stateV2 = stateV2.SetState(
-                    PetState.DeriveAddress(_avatarAddr, (int)_petId),
-                    new List(_petId!.Serialize(), petLevel.Serialize(), 0L.Serialize())
-                );
-                expectedBlock = (int)Math.Round(
-                    expectedBlock * (1 - petRow.LevelOptionMap[(int)petLevel].OptionValue / 100)
-                );
-            }
-
-            // Prepare
-            stateV2 = CraftUtil.UnlockStage(
-                stateV2,
-                _tableSheets,
-                _worldInfoAddr,
-                GameConfig.RequireClearedStageLevel.ItemEnhancementAction
-            );
-            stateV2 = CraftUtil.PrepareCombinationSlot(stateV2, _avatarAddr, 0);
-
-            // Do Enhancement
-            var action = new ItemEnhancement
-            {
-                avatarAddress = _avatarAddr,
-                itemId = equipment.ItemId,
-                materialId = material.ItemId,
-                slotIndex = 0,
-                petId = _petId,
-            };
-
-            stateV2 = action.Execute(new ActionContext
-            {
-                PreviousStates = stateV2,
-                Signer = _agentAddr,
-                BlockIndex = 0L,
-                Random = random,
-            });
-
-            var slotState = stateV2.GetCombinationSlotState(_avatarAddr, 0);
-            // TEST: RequiredBlockIndex
-            Assert.Equal(expectedBlock, slotState.RequiredBlockIndex);
-        }
     }
 }
