+++ Lib9c/Action/AttachmentActionResult.cs
@@ -21,7 +21,7 @@ namespace Nekoyume.Action
                 ["combination.result-model"] = d => new CombinationConsumable5.ResultModel(d),
                 ["itemEnhancement.result"] = d => new ItemEnhancement7.ResultModel(d),
                 ["item_enhancement9.result"] = d => new ItemEnhancement9.ResultModel(d),
-                ["item_enhancement11.result"] = d => new ItemEnhancement11.ResultModel(d),
+                ["item_enhancement11.result"] = d => new ItemEnhancement.ResultModel(d),
                 ["sellCancellation.result"] = d => new SellCancellation.Result(d),
                 ["rapidCombination.result"] = d => new RapidCombination0.ResultModel(d),
                 ["rapid_combination5.result"] = d => new RapidCombination5.ResultModel(d),
