+++ Lib9c/Extensions/CombinationSlotStateExtensions.cs
@@ -59,9 +59,6 @@ namespace Nekoyume.Extensions
                 case ItemEnhancement10.ResultModel r:
                     resultId = r.id;
                     break;
-                case ItemEnhancement11.ResultModel r:
-                    resultId = r.id;
-                    break;
                 case MonsterCollectionResult r:
                     resultId = r.id;
                     break;
@@ -108,7 +105,7 @@ namespace Nekoyume.Extensions
                 case DailyReward2.DailyRewardResult r:
                     resultId = r.id;
                     break;
-                case ItemEnhancement11.ResultModel r:
+                case ItemEnhancement.ResultModel r:
                     resultId = r.id;
                     break;
                 case ItemEnhancement7.ResultModel r:
@@ -185,13 +182,6 @@ namespace Nekoyume.Extensions
                         resultId,
                         requiredBlockIndex);
                     return true;
-                case ItemEnhancement11.ResultModel r:
-                    itemEnhanceMail = new ItemEnhanceMail(
-                        r,
-                        blockIndex,
-                        resultId,
-                        requiredBlockIndex);
-                    return true;
                 case CombinationConsumable5.ResultModel r:
                     combinationMail = new CombinationMail(
                         r,
@@ -222,7 +212,7 @@ namespace Nekoyume.Extensions
 
             switch (state.Result)
             {
-                case ItemEnhancement11.ResultModel r:
+                case ItemEnhancement.ResultModel r:
                     itemEnhanceMail = new ItemEnhanceMail(
                         r,
                         blockIndex,
